16. Parsing Variables from URLs
ref:  https://www.udemy.com/course/nodejs-express-mongodb-bootcamp/learn/lecture/15080942

1.    Modified code section:    // === PRODUCT page === //
      (Single page for each product clicked on "overview" page inside
      the section:
      const server = http.createServer((req, res) => {      )

2.    // decide which property to be used for rendering the page

    const productObjContent = dataObj[query.id];
    // dataObj is from JSON file (Array format),
    // [query.id] is the obj assigned with the obj from the method url.parse(req.url, true)

3. process template-product.html and replace key words with the value inside the obj property

### For res.end()
ref:
request.end([data[, encoding]][, callback]) method:
https://nodejs.org/api/http.html#http_request_end_data_encoding_callback
Finishes sending the request. If any parts of the body are unsent, it will flush them to the stream. If the request is chunked, this will send the terminating '0\r\n\r\n'.

If data is specified, it is equivalent to calling request.write(data, encoding) followed by request.end(callback).

If callback is specified, it will be called when the request stream is finished.

https://stackoverflow.com/questions/29555290/what-is-the-difference-between-res-end-and-res-send

res.end() will end the response process. This method actually comes from Node core, specifically the response.end() method of http.ServerResponse. It is used to quickly end the response without any data. For example:

res.end();
res.status(404).end();